:wrench:fix: publish product detail modal without fetch
